#49 move PageBatchEvent from infrastructure to domain package and rename PageBatchEventAdapter to RepositoryChangedEventAdapter
