+++ src/main/java/com/cloudogu/smeagol/wiki/domain/ChangeType.java
@@ -1,4 +1,4 @@
-package com.cloudogu.smeagol.wiki.infrastructure;
+package com.cloudogu.smeagol.wiki.domain;
 
 /**
  * Type of a change.
